MLK-17962-3 core: arm: imx: tune PL310 configuration
The current tag/data ram, prefetch value is not the best value.
With this, the performance is not good. So retune the value
to match i.MX design to have good performance.

Also there is PL310 errata that we need to disable Double linefill
for version below r3p2 pl310.

Signed-off-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Alessandro Di Chiara <alessandro.dichiara@nxp.com>