MLK-18036-2 Add dt_find_ocram_tz_addr() function
The OCRAM space allocated to optee is Trustzone protected. Only the
begin address of the protection is specifiable.
This function fetches the address of the node to protect in the dtb.
Nodes to protect are specified by its compatible string in
tz_ocram_match[] tab.

This function defines the beginning the start address of the TZ ocram
space.

Signed-off-by: Clement Faure <clement.faure@nxp.com>
(cherry picked from commit 5b29889537ef8c6d912d1fdb33fe308a295b2b60)