TEE-301: Protect memory against aliased access
The TZASC must be configured to handle aliased addresses if
the DDR installed does not occupy all the memory space
mappable by the controller as expalined in the RM.

In order to protect against the access of secure data using
an aliased address which is not secured, we set all the
memory space mappable as secure. This way all aliased accesses
will be treated as needed secure access.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
(cherry picked from commit ab91a1fd4196d300e455a001175bcb2214005b5c)