MLK-18981: optee memory overlaps with m4 firmware
when M4 is being used on i.MX6SX and 7D, the firmware uses the same memory
 region as optee firmware.
Harmonized all armv7 SoC to use the same location.
Moved optee ddr to an offset of 64MB from the start of DDR.
optee updates the device tree to let Linux kernel knows not
to use this region.
TZASC settings are updated accordingly to protect this memory.

Signed-off-by: Silvano di Ninno <silvano.dininno@nxp.com>