MLK-19654 Reduce size allocated to OCRAM suspend code
On imx6sx, power management code and TLB are both allocated in the OCRAM_S.
Due to the small size of the OCRAM_S on this board, TLB and suspend code
overlaps. After a previous code maintenance, the cpu idle code overwrites
first TLB entries. To fix this, the suspend code space allocation is
decreased and the cpu idle code offset is lowered.

Signed-off-by: Clement Faure <clement.faure@nxp.com>