MLK-21281: Save CAAM JR MID for the JR not used by TEE
The code was saving the MIDs for the JR if CFG_IMXCRYPT was set.

In case of suspend, the MID of the JR is resetted especially the JROWN_NS
bit which give access to the JR from Non-Secure. The MID is not modified
before the NS OS resume and the JRs are no more accessible.

This patch move the logic to do the backup to another function which
is only called from hal_cfg_setup_nsjobring (hal_cfg.c) which is
only called once.

The JR shared by the TEE and NS OS has its owner configured by the TEE
before the NS OS resume.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>