TEE-482 Add and remove clang-format rules
From Kernel documentation:
"clang-format is a tool to format C/C++/… code according to a set
of rules and heuristics. Like most tools, it is not perfect nor
covers every single case, but it is good enough to be helpful"

Even if clang-format is described as a helpful development tool,
we will use it as a code style enforcer in OPTEE project. We allow
some flexibility by adding/removing rules.

Add AlignConsecutiveMacros rule and set it to true.
Remove AlignConsecutiveAssignments rule.
Remove AlignTrailingComments rule.

Since the AlignConsecutiveMacros rule is added, Clang 9.0 or above
is required. To download Clang 9.0 or above:
https://releases.llvm.org/download.html

For more info about clang-format rules:
https://clang.llvm.org/docs/ClangFormatStyleOptions.html

Signed-off-by: Clement Faure <clement.faure@nxp.com>