TEE-487 Tweak PenaltyBreakBeforeFirstCallParameter value
In some cases, clang-format would break long function calls like this:
---------------------------
functionCall(
	arg1, arg2, arg3)
---------------------------
instead of this:
---------------------------
functionCall(arg1,
	     arg2, arg3)
---------------------------

The formatting proposed by Clang-format did not comply the Checkpatch
rule : CHECK: Lines should not end with a '('

To enforce this rule, clang-format penalty values can be tweaked to
discourage some formating behaviour. In this case the value to tweak is
PenaltyBreakBeforeFirstCallParameter. By setting this value to 80
instead of 30, we suggest Clang-format to always break line after the
first parameter below character limit of 80.

Signed-off-by: Clement Faure <clement.faure@nxp.com>