TEE-529 core: imx: Release SM partition after DEK blob creation
The Blob SM driver expects the caller to free the Secure Memory
partition after DEK Blob creation.
U-boot does not do it, because it was not required to do so
in the imx_3.2.y implementation. To keep backward compatibility
with older optee-os releases do the free in the pta.

Signed-off-by: Silvano di Ninno <silvano.dininno@nxp.com>
Fixes: 4fcee7333abd ("TEE-517 core: imx: Fix DEK Blob support for i.MX 8mscale family")
(cherry picked from commit a7698797984a2a0c3f627caa4ced982619c07dc7)