TEE-579 Fix DCP context buffer on imx6sllevk
In the DCP driver, the CTRL register was configured without CTRL_SET/CLR
sub-registers. This believes to lead to an undefined behaviour since the
same DCP driver was working fine on imx6ull/ulz but not on imx6sll.

This commit fixes this issue and add CTRL_SET/CLR registers to properly:
 * Set DCP_CTRL_ENABLE_CONTEXT_SWITCHING and
DCP_CTRL_GATHER_RESIDUAL_WRITES bits in CTRL register.
 * Clear DCP_CTRL_SFTRST and DCP_CTRL_CLKGATE bits in CTRL register.

It's also important to set the DCP_CONTEXT register after the DCP_CTRL
register. Otherwise, the DCP_CONTEXT might be wiped.

Signed-off-by: Clement Faure <clement.faure@nxp.com>