TEE-595 core: driver: remove ECC P521 key size check
This patch removes the key size check for ECC P521 generation. Because
of the secure key and the public key being allocated in the same buffer
in the CAAM driver, the key size shall be byte aligned.
In this case, to generate a 521 bits ECC key, a buffer of 528 bits is
allocated.

Signed-off-by: Clement Faure <clement.faure@nxp.com>