LFOPTEE-3 drivers: caam: Fix warning of caam_desc_pop
When caam driver is built with CFG_CAAM_64BIT the following warning
happens:
| core/drivers/crypto/caam/caam_desc.c: In function 'caam_desc_pop':
	core/drivers/crypto/caam/caam_desc.c:83:43: error:
	dereferencing type-punned pointer will break strict-aliasing rules
| [-Werror=strict-aliasing]
|    83 |  const uint32_t *a32 = (const uint32_t *)(&out_entry->desc);

This issue happens because &out_entry->desc is a pointer on a 64 value
casted to a pointer on a 32bit value.

The issue is fixed by using an intermediate value to store the value
of the pointer before assigning it to the 32bit pointer.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>