LFOTPEE-20 drivers: caam: fix HMAC key context copy
When copying HMAC context, the content of the key buffer was copied
into the new HMAC context but not the key length.
This caused the new HMAC context to not use the HMAC key for its
operations, resulting on a wrong result.

Signed-off-by: Clement Faure <clement.faure@nxp.com>
(cherry picked from commit 7325acc5afa933482221381e00fdd5f6bbb2d252)