LFOPTEE-21 drivers: caam: fix missing output additional length
CAAM HW Module requires to have additional bytes in output buffer
to perform the requested operation.
This is the case for ECDSA operation, signature must be 16 bytes
aligned.
The CAAM DMA object is adding the required bytes if the signature
output buffer is too short. The issue is that this additional
bytes were not added in the DMA buffer to be allocated.

Signed-off-by: Cedric Neveux <cedric.neveux@nxp.com>