LFOPTEE-39 pta: dek_blob: Wrong size for the blob used
Form the buffer passed as input, a part is used to store the
DEK blob header, the rest is for the blob itself.

When setting the buffer for the creation of the blob, the pointer
is advanced by the size of the debblob header. The size is also
wrongly increased.

The issue can create a memory corruption.

This patch fixes the issue by decreasing the size for the buffer
instead of increasing it.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>