LF-3825 drivers: caam: ack the interrupt when completing a job
The ITR will only be acked if the job finishes between the
do_jr_dequeue() and caam_hal_jr_check_ack_itr().

With the patch, we are sure the ITR is acked when we are
leaving the function.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
(cherry picked from commit 671a1648e340735ea443b805fd6a10cb4fe0b663)