LFOPTEE-66 core: imx: compile pm-imx8.c for all 8mscale platforms
Enable the compilation of pm-imx8.c file for all 8mscale platforms.
Before, this file was only compiled for mx8mm.

This file is need to trigger the OPTEE power management callbacks upon
suspend and resume operations. This is needed for some drivers that
require power management operations.

Introduce a CFG_MX8M flag that is enabled for all 8mscale platforms.

Signed-off-by: Clement Faure <clement.faure@nxp.com>