LFOPTEE-89 drivers: caam: Fix cache operation on SGT table
The cache operation on SGT table needs to take into account the
"burst" required by the CAAM implemented by the value
CFG_CAAM_SGT_ALIGN.

One cache operation on SGT table was not implemented with this
consideration.

This patch fix the issue by encapsulating the cache the operation
on SGT table to avoid implementation errors.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>