drivers: pm: remove file with license issue
The removed file has a copyright/license issue.

A blackduck scan reveals this file has a a dual license
"GPL-2.0 OR BSD-2-Clause" but the source file in the Linux kernel
is "GPL-2.0-only".

This file has been merged in the 3.17.0 optee-os upstream release.

As a precaution, this file will be removed from NXP tree. The upstream
community has been informed of the issue.

Signed-off-by: Clement Faure <clement.faure@nxp.com>