LFOPTEE-169 imx_ele: use get_info command to get the soc revision
Use the ELE get info command to get the SoC revision.
This command will fail on A0 SoC revision as this command is not
implemented on this SoC revision. In this case, set the revision to A0.

Some ELE commands are only available on 8ULP A1 SoC revision.
Read JTAG_ID_REG register to fetch the SoC revision.

Signed-off-by: Clement Faure <clement.faure@nxp.com>