LF-7525 drivers: dcp: do not modify DCP node status in the DTB
The DCP driver in OPTEE sets "status" to "disabled" and "secure-status"
to "enabled". This prevents the DCP driver in Linux from probing.

In OPTEE, the DCP is only used to generate the HUK at boot. There is no
need for OPTEE to prevent the Linux kernel from probing the DCP and
using it at runtime.

Signed-off-by: Clement Faure <clement.faure@nxp.com>